# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "小红书爆款笔记生成 agent Skill",
    "description": "Enterprise-grade hot-note research, multi-model generation, account analytics and controlled publishing Agent Skill.",
    "version": "5.14.5"
  },
  "paths": {
...
    "/v1/assets": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Upload Asset",
        "operationId": "upload_asset_v1_assets_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_asset_v1_assets_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Upload Asset V1 Assets Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/operations/accounts/{account_id}/notes/{note_id}/attribution": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Performance Attribution",
        "operationId": "performance_attribution_v1_operations_accounts__account_id__notes__note_id__attribution_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Performance Attribution V1 Operations Accounts  Account Id  Notes  Note Id  Attribution Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/operations/accounts/{account_id}/weight-trend": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Weight Trend",
        "operationId": "weight_trend_v1_operations_accounts__account_id__weight_trend_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Weight Trend V1 Operations Accounts  Account Id  Weight Trend Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/operations/calendar": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Create Calendar",
        "operationId": "create_calendar_v1_operations_calendar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Calendar V1 Operations Calendar Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/operations/series": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Create Series",
        "operationId": "create_series_v1_operations_series_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Series V1 Operations Series Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/operations/experiments": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Create Experiment",
        "operationId": "create_experiment_v1_operations_experiments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Experiment"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Create Experiment V1 Operations Experiments Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/operations/experiments/{experiment_id}/assign": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Assign Experiment",
        "operationId": "assign_experiment_v1_operations_experiments__experiment_id__assign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Assign Experiment V1 Operations Experiments  Experiment Id  Assign Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/operations/experiments/outcomes": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Record Experiment Outcome",
        "operationId": "record_experiment_outcome_v1_operations_experiments_outcomes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentOutcome"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Record Experiment Outcome V1 Operations Experiments Outcomes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/operations/experiments/{experiment_id}/analysis": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Analyze Experiment",
        "operationId": "analyze_experiment_v1_operations_experiments__experiment_id__analysis_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "minimum_samples_per_variant",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Minimum Samples Per Variant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Analyze Experiment V1 Operations Experiments  Experiment Id  Analysis Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/operations/bandit/choose": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Choose Bandit",
        "operationId": "choose_bandit_v1_operations_bandit_choose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanditDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Choose Bandit V1 Operations Bandit Choose Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/operations/bandit/update": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Update Bandit",
        "operationId": "update_bandit_v1_operations_bandit_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanditUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Update Bandit V1 Operations Bandit Update Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/operations/assets/import": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Import Asset",
        "operationId": "import_asset_v1_operations_assets_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Import Asset V1 Operations Assets Import Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/operations/assets": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Search Assets",
        "operationId": "search_assets_v1_operations_assets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Assets V1 Operations Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/operations/accounts/{account_id}/notes/{note_id}/retrospective": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Retrospective",
        "operationId": "retrospective_v1_operations_accounts__account_id__notes__note_id__retrospective_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Retrospective V1 Operations Accounts  Account Id  Notes  Note Id  Retrospective Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignmentRequest": {
        "properties": {
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          }
        },
        "type": "object",
        "required": [
          "subject_id"
        ],
        "title": "AssignmentRequest"
      },
...
      "BanditDecisionRequest": {
        "properties": {
          "policy_id": {
            "type": "string",
            "title": "Policy Id"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "arms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Arms"
          },
          "context": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 128,
            "title": "Context"
          },
          "features": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Features"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "auto_account_weight": {
            "type": "boolean",
            "title": "Auto Account Weight",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "policy_id",
          "subject_id",
          "arms"
        ],
        "title": "BanditDecisionRequest"
      },
...
      "Body_upload_asset_v1_assets_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_asset_v1_assets_post"
      },
      "CalendarRequest": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Topics"
          },
          "fallback_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Fallback Topics"
          },
          "days": {
            "type": "integer",
            "maximum": 365.0,
            "minimum": 7.0,
            "title": "Days",
            "default": 30
          },
          "posts_per_week": {
            "type": "integer",
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Posts Per Week",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "account_id"
        ],
        "title": "CalendarRequest"
      },
...
      "Experiment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "default": "local"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "hypothesis": {
            "type": "string",
            "title": "Hypothesis"
          },
          "primary_metric": {
            "type": "string",
            "title": "Primary Metric"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/ExperimentVariant"
            },
            "type": "array",
            "title": "Variants"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "DRAFT"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "name",
          "hypothesis",
          "primary_metric",
          "variants"
        ],
        "title": "Experiment"
      },
      "ExperimentOutcome": {
        "properties": {
          "experiment_id": {
            "type": "string",
            "title": "Experiment Id"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "variant_id": {
            "type": "string",
            "title": "Variant Id"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          }
        },
        "type": "object",
        "required": [
          "experiment_id",
          "subject_id",
          "variant_id",
          "metric",
          "value"
        ],
        "title": "ExperimentOutcome"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SeriesRequest": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "audience": {
            "type": "string",
            "title": "Audience"
          },
          "episode_count": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 2.0,
            "title": "Episode Count",
            "default": 6
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "title",
          "topic",
          "audience"
        ],
        "title": "SeriesRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}